Moved all existing enums to lib for reusability
